# Add AVX optimizations for 7-tuple and 8-tuple sieving