--- a/Miner.hpp
+++ b/Miner.hpp
@@ -17,7 +17,7 @@ union xmmreg_t {
 	__m128i m128;
 };
 
-constexpr uint32_t sieveCacheSize(16);
+constexpr uint32_t sieveCacheSize(32);
 constexpr uint32_t nWorks(2);
 
 inline mpz_class u64ToMpz(const uint64_t u64) {
@@ -147,6 +147,7 @@ class Miner {
 	void _doPresieveTask(const Task&);
 	void _processSieve(uint64_t*, uint32_t*, const uint64_t, const uint64_t);
 	void _processSieve6(uint64_t*, uint32_t*, uint64_t, const uint64_t);
+	void _processSieve7(uint64_t*, uint32_t*, uint64_t, const uint64_t);
 	void _doSieveTask(Task);
 	bool _testPrimesIspc(const std::array<uint32_t, maxCandidatesPerCheckTask>&, uint32_t[maxCandidatesPerCheckTask], const mpz_class&, mpz_class&);
 	void _doCheckTask(Task);