--- a/Miner.hpp
+++ b/Miner.hpp
@@ -5,6 +5,7 @@
 
 #include <atomic>
 #include <cassert>
+#include <immintrin.h>
 #include "Stats.hpp"
 #include "Client.hpp"
 #include "StratumClient.hpp"
@@ -15,6 +16,12 @@ union xmmreg_t {
 	uint32_t v[4];
 	uint64_t v64[2];
 	__m128i m128;
+};
+
+union ymmreg_t {
+	uint32_t v[8];
+	uint64_t v64[4];
+	__m256i m256;
 };
 
 constexpr uint32_t sieveCacheSize(32);
@@ -148,6 +155,8 @@ class Miner {
 	void _processSieve(uint64_t*, uint32_t*, const uint64_t, const uint64_t);
 	void _processSieve6(uint64_t*, uint32_t*, uint64_t, const uint64_t);
 	void _processSieve7(uint64_t*, uint32_t*, uint64_t, const uint64_t);
+	void _processSieve8(uint64_t*, uint32_t*, uint64_t, const uint64_t);
+	void _processSieve8_avx2(uint64_t*, uint32_t*, uint64_t, const uint64_t);
 	void _doSieveTask(Task);
 	bool _testPrimesIspc(const std::array<uint32_t, maxCandidatesPerCheckTask>&, uint32_t[maxCandidatesPerCheckTask], const mpz_class&, mpz_class&);
 	void _doCheckTask(Task);