Added AVX and AVX2 versions of process sieve for 7 and 8 tuples.