@Pttn 2021-04-08
Sorry for the delay, I was focused on the Riecoin 0.21 upgrade that I will reveal soon (well I know that at least one person has seen it Lol). I actually mined several days with your code, the speedups seem great (+2-4% without AVX2 and +10-15% with it enabled at Current Difficulties, compared to previous non-AVX2/AVX2 code respectively; even more for 8-tuples), but I noticed that I actually found very few blocks.

I did some tests in Benchmark Mode to make sure that I was not just very unlucky and the results seem to confirm that there is a bug in your Sieving code, could you check it?
Note that I did not check for this bug for 8-tuples nor without AVX2.

At Difficulty 256, PrimeTableLimit 262144, 5 minutes Benchmark, 3700X 16 Threads/8 SieveWorkers, no Block, AVX2 Enabled
Without Optimizations
462019.703208 candidates/s, ratio 7.982996 -> 19320.259839 block(s)/day
Tuples found: (138606288 17362691 2172228 271522 34018 4297 569 79) in 300.000846 s
Tuple rates : (462019.656743 57875.473412 7240.739575 905.070780 113.393013 14.323293 1.896661 0.263333)
Tuple ratios: (7.982996 7.993033 8.000192 7.981716 7.916686 7.551845 7.202532)

With Optimizations
537697.914097 candidates/s, ratio 7.982295 -> 22498.707143 block(s)/day
Tuples found: (161311827 20208702 2530446 316536 39424 5063 647 56) in 300.004590 s
Tuple rates : (537697.862710 67361.309308 8434.690941 1055.103856 131.411323 16.876408 2.156634 0.186664)
Tuple ratios: (7.982295 7.986221 7.994181 8.029018 7.786688 7.825348 11.553571)

You can see that the last Ratio 11.553571 is pretty bad and cannot be explained by the Luck Factor.
The Benchmark Mode is fairly deterministic in different runs with same parameters, the same tuples would be found again. Many tuples found without optimizations do not appear when running with the new code.
I tested shortly a few more Difficulties and reached the same conclusions.

Otherwise I would really love to merge it, and hopefully the fixes do not reduce much the speedup.