@MichaelBell 2021-04-08
Thank you for your diligent testing.  I had missed this - incorrectly putting it down to bad luck.  I understand the problem now.  It only affects 7-tuples on AVX2, the other modes should be OK.  I'll have a think about how best to fix it.