@MichaelBell 2021-04-09
I believe this is now correct and performance should be basically the same.