@Pttn 2021-04-09
Seems Ok to me, I tested the four combinations 7/8-tuples x non-AVX2/AVX2 without issues. Great job, thank you!

Will you also make a Pull Request for the custom GMP code you talked about on Discord?
Does it still provide good speedups for 7-tuples as Prime Testing now takes less time?