@MichaelBell 2021-04-09
Thanks.

I need to go back to that custom GMP code and work out what is wrong with it - it is no longer working on machines that it used to work on - I guess due to some header inconsistency with an updated the version of GMP from the OS or something.  But as you say, the speedup won't be as great now anyway.